REmoved console statements

+++ Backend/routes/profile.js
@@ -61,15 +61,19 @@ router.put('/update',async function(req,res,next){
 
 
 // Which controller?
-router.delete('/:id', function(req,res, next){
+router.delete('/', function(req,res, next){
     if(req.body.username==null){
         res.status(422);
         res.send({error: "no username provided"});
     } else {
         User.findOne({username : req.body.username}, {_id : 1}).then(function(user_id){
-            User.findByIdAndDelete({_id : user_id._id}).then(function(user){
-                res.send(user);
-            }).catch(next);
+            if(user_id == null){
+                errorController.error(res, "user not found", 400);
+            } else {
+                User.findByIdAndDelete({_id : user_id._id}).then(function(user){
+                    res.send(user);
+                }).catch(next);
+            }            
         }).catch(next);
     }
 });
